pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arkana-mdd"
version = "0.1.0"
description = "Metadata-driven data selection framework with FastAPI runtime"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "keyring>=25.0.0",
  "mysql-connector-python>=9.0.0",
  "pydantic>=2.9.0",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.30.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
